Pre-final version

+++ website/private/System/Handlers/MiscellaneousHandler.php
@@ -6,7 +6,9 @@ namespace System\Handlers;
 use System\Core\Domain\DTO\ResponseDTO;
 use System\Core\Exceptions\InvalidArgumentException;
 use System\Core\Exceptions\NotFoundException;
+use System\Core\Exceptions\NotLoggedInException;
 use System\Miscellaneous\MiscController;
+use System\Router;
 
 class MiscellaneousHandler {
 
@@ -21,6 +23,7 @@ class MiscellaneousHandler {
             case "provinces": return $this->processProvinces($action, $data);
             case "cities": return $this->processCities($action, $data);
             case "zipcodes": return $this->processZipcodes($action, $data);
+            case "address": return $this->processAddresses($action, $data);
             case "categories": return $this->processCategories($action, $data);
             case "subcategories": return $this->processSubCategories($action, $data);
             case "questions": return $this->processQuestions($action, $data);
@@ -90,6 +93,8 @@ class MiscellaneousHandler {
     }
     
     private function newQuestion(string $productId, string $message): bool {
+        if(Router::$CURRENT_USER === null) throw new NotLoggedInException();
+
         $errors = Array();
 
         if(empty($productId) || !is_numeric($productId)) $errors[] = "PRODUCT_ID_MUST_BE_INT";
@@ -100,11 +105,15 @@ class MiscellaneousHandler {
     }
 
     private function deleteQuestion(string $questionId): bool {
+        if(Router::$CURRENT_USER === null) throw new NotLoggedInException();
+        
         if(empty($questionId) || !is_numeric($questionId)) throw new InvalidArgumentException("QUESTION_ID_MUST_BE_INT");
         return $this->controller->deleteQuestion(intval($questionId));
     }
 
     private function updateQuestion(string $questionId, string $message): bool {
+        if(Router::$CURRENT_USER === null) throw new NotLoggedInException();
+
         $errors = Array();
 
         if(empty($questionId) || !is_numeric($questionId)) $errors[] = "QUESTION_ID_MUST_BE_INT";
@@ -115,6 +124,8 @@ class MiscellaneousHandler {
     }
 
     private function respondTo(string $questionId, string $response): bool {
+        if(Router::$CURRENT_USER === null) throw new NotLoggedInException();
+
         $errors = Array();
 
         if(empty($questionId) || !is_numeric($questionId)) $errors[] = "QUESTION_ID_MUST_BE_INT";
@@ -125,11 +136,15 @@ class MiscellaneousHandler {
     }
 
     private function deleteResponse(string $questionId): bool {
+        if(Router::$CURRENT_USER === null) throw new NotLoggedInException();
+
         if(empty($questionId) || !is_numeric($questionId)) throw new InvalidArgumentException("QUESTION_ID_MUST_BE_INT");
         return $this->controller->deleteResponse(intval($questionId));
     }
 
     private function updateResponse(string $questionId, string $message): bool {
+        if(Router::$CURRENT_USER === null) throw new NotLoggedInException();
+
         $errors = Array();
 
         if(empty($questionId) || !is_numeric($questionId)) $errors[] = "QUESTION_ID_MUST_BE_INT";
@@ -140,6 +155,8 @@ class MiscellaneousHandler {
     }
 
     private function createReview(string $productId, string $stars, string $title, string $description): bool {
+        if(Router::$CURRENT_USER === null) throw new NotLoggedInException();
+
         $errors = Array();
 
         if(empty($productId) || !is_numeric($productId)) $errors[] = "PRODUCT_ID_MUST_BE_INT";
@@ -153,11 +170,15 @@ class MiscellaneousHandler {
     }
 
     private function deleteReview(string $reviewId): bool {
+        if(Router::$CURRENT_USER === null) throw new NotLoggedInException();
+
         if(empty($reviewId) || !is_numeric($reviewId)) throw new InvalidArgumentException("REVIEW_ID_MUST_BE_INT");
         return$this->controller->deleteReview(intval($reviewId));
     }
 
     private function updateReview(string $reviewId, string $stars, string $title, string $description): bool {
+        if(Router::$CURRENT_USER === null) throw new NotLoggedInException();
+
         $errors = Array();
         if(empty($reviewId) || !is_numeric($reviewId)) $errors[] = "REVIEW_ID_MUST_BE_INT";
         if(empty($stars) || !is_numeric($stars)) $errors[] = "STARS_MUST_BE_INT";
@@ -205,16 +226,61 @@ class MiscellaneousHandler {
 
     private function processZipcodes(string $action, Array $data): ResponseDTO {
         switch($action) {
-            case "verify": return $this->verifyZipcode($data['zip-code'] ?? "", $data['province_id'] ?? "");
+            case "verify": return new ResponseDTO($this->verifyZipcode($data['zip-code'] ?? "", $data['province_id'] ?? ""));
             default: throw new NotFoundException();
         }
     }
 
-    private function verifyZipcode(string $zipcode, $province_id): ResponseDTO {
+    private function processAddresses(string $action, Array $data): ResponseDTO {
+        switch($action) {
+            case "get": return new ResponseDTO($this->getUserAddresses());
+            case "delete": return new ResponseDTO($this->deleteAddress($data['address_id'] ?? ""));
+            case "update": return new ResponseDTO(
+                    $this->updateAddress(
+                        $data['address_id'] ?? "", 
+                        $data['province_id'] ?? "", 
+                        $data['city_id'] ?? "", 
+                        $data['zip_code'] ?? "", 
+                        $data['street'] ?? "", 
+                        $data['number'] ?? "", 
+                        $data['extra'] ?? "", 
+                    )
+                );
+            case "create": return new ResponseDTO(
+                    $this->createAddress(
+                        $data['province_id'] ?? "", 
+                        $data['city_id'] ?? "", 
+                        $data['zip_code'] ?? "", 
+                        $data['street'] ?? "", 
+                        $data['number'] ?? "", 
+                        $data['extra'] ?? "", 
+                    )
+                );
+            default: throw new NotFoundException();
+        }
+    }
+
+    private function getUserAddresses(): Array {
+        if(Router::$CURRENT_USER === null) throw new NotLoggedInException();
+
+        return $this->controller->getAddressesByUserId(Router::$CURRENT_USER->getId());
+    }
+
+    private function deleteAddress(string $addressId): bool {
+        if(Router::$CURRENT_USER === null) throw new NotLoggedInException();
+        if(empty($addressId) || !is_numeric($addressId)) throw new InvalidArgumentException("INVALID_ADDRESS_ID");
+
+        $address = $this->controller->getAddressById(intval($addressId));
+        if(!$address) throw new NotFoundException("ADDRESS_ID");
+
+        return $this->controller->deleteAddress($address);
+    }
+
+    private function verifyZipcode(string $zipcode, $province_id): bool {
         if(empty($zipcode) || !is_numeric($zipcode)) throw new InvalidArgumentException(Array("ZIPCODE_MUST_BE_INT"));
         if(empty($province_id) || !is_numeric($province_id)) throw new InvalidArgumentException(Array("PROVINCE_ID_MUST_BE_INT"));
         $valid = $this->controller->verifyZipcode(intval($zipcode), intval($province_id));
-        if($valid) return new ResponseDTO(true);
+        if($valid) return true;
         else throw new InvalidArgumentException(Array("ZIPCODE_DOES_NOT_MATCH_PROVINCE"));
     }
     
@@ -230,12 +296,16 @@ class MiscellaneousHandler {
     }
 
     private function deleteCategory(string $id): bool {
+        if(Router::$CURRENT_USER === null) throw new NotLoggedInException();
+
         if(empty($id) || !is_numeric($id)) throw new InvalidArgumentException(Array("ID_MUST_BE_INT"));
 
         return $this->controller->deleteCategory(intval($id));
     }
 
     private function updateCategory(string $name, $id): bool {
+        if(Router::$CURRENT_USER === null) throw new NotLoggedInException();
+
         if(empty($id) || !is_numeric($id)) throw new InvalidArgumentException(Array("ID_MUST_BE_INT"));
         if(empty($name)) throw new InvalidArgumentException(Array("NAME_CANNOT_BE_EMPTY"));
 
@@ -243,6 +313,8 @@ class MiscellaneousHandler {
     }
 
     private function newSubCategory(string $name, string $categoryId): bool {
+        if(Router::$CURRENT_USER === null) throw new NotLoggedInException();
+
         if(empty($categoryId) || !is_numeric($categoryId)) throw new InvalidArgumentException(Array("ID_MUST_BE_INT"));
         if(empty($name)) throw new InvalidArgumentException(Array("NAME_CANNOT_BE_EMPTY"));
         return $this->controller->newSubCategory(htmlentities($name), intval($categoryId));
@@ -254,6 +326,8 @@ class MiscellaneousHandler {
     }
 
     private function deleteSubCategory(string $id, string $categoryId): bool {
+        if(Router::$CURRENT_USER === null) throw new NotLoggedInException();
+
         if(empty($id) || !is_numeric($id)) throw new InvalidArgumentException(Array("ID_MUST_BE_INT"));
         if(empty($categoryId) || !is_numeric($categoryId)) throw new InvalidArgumentException(Array("CATEGORY_ID_MUST_BE_INT"));
 
@@ -261,10 +335,46 @@ class MiscellaneousHandler {
     }
 
     private function updateSubCategory(string $name, string $id, string $categoryId): bool {
+        if(Router::$CURRENT_USER === null) throw new NotLoggedInException();
+
         if(empty($id) || !is_numeric($id)) throw new InvalidArgumentException(Array("ID_MUST_BE_INT"));
         if(empty($categoryId) || !is_numeric($categoryId)) throw new InvalidArgumentException(Array("CATEGORY_ID_MUST_BE_INT"));
         if(empty($name)) throw new InvalidArgumentException(Array("NAME_CANNOT_BE_EMPTY"));
 
         return $this->controller->updateSubCategory(htmlentities($name), intval($id), intval($categoryId));
     }
+
+    private function updateAddress(string $addressId, string $provinceId, string $cityId, string $zipCode, string $street, string $number, string $extra): bool {
+        if(Router::$CURRENT_USER === null) throw new NotLoggedInException();
+
+        $errors = Array();
+
+        if(empty($addressId) || !is_numeric($addressId)) $errors['INVALID_ADDRESS_ID'];
+        if(empty($provinceId) || !is_numeric($provinceId)) $errors['INVALID_PROVINCE_ID'];
+        if(empty($cityId) || !is_numeric($cityId)) $errors['INVALID_CITY_ID'];
+        if(empty($zipCode) || !is_numeric($zipCode)) $errors['INVALID_ZIP_CODE'];
+        if(empty($street)) $errors['EMPTY_STREET'];
+        if(empty($number) || !is_numeric($number)) $errors['INVALID_STREET_NUMBER'];
+
+        if(count($errors) != 0) throw new InvalidArgumentException($errors);
+
+        return $this->controller->updateAddress(intval($addressId), intval($provinceId), intval($cityId), intval($zipCode), $street, intval($number), $extra);
+    }
+
+    private function createAddress(string $provinceId, string $cityId, string $zipCode, string $street, string $number, string $extra): bool | int {
+        if(Router::$CURRENT_USER === null) throw new NotLoggedInException();
+
+        $errors = Array();
+
+        if(empty($provinceId) || !is_numeric($provinceId)) $errors['INVALID_PROVINCE_ID'];
+        if(empty($zipCode) || !is_numeric($zipCode)) $errors['INVALID_ZIP_CODE'];
+        if(empty($cityId) || !is_numeric($cityId)) $errors['INVALID_CITY_ID'];
+        if(empty($street)) $errors['EMPTY_STREET'];
+        if(empty($number) || !is_numeric($number)) $errors['INVALID_STREET_NUMBER'];
+
+        if(count($errors) != 0) throw new InvalidArgumentException($errors);
+
+        return $this->controller->createAddress(intval($provinceId), intval($cityId), intval($zipCode), $street, intval($number), $extra);
+    }
+    
 }
